NullReferenceCheck added to RecentFileList fixes #851
git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@6057 9b0d660e-8ab2-dd11-ad6d-000c29d71a56

+++ Startcenter/LastOpenedFilesList.xaml.cs
@@ -169,6 +169,10 @@ namespace Startcenter
 
         private void MenuItem_Click_1(object sender, RoutedEventArgs e)
         {
+            if (RecentFileListBox == null || RecentFileListBox.SelectedItem == null)
+            {
+                return;
+            }
             var file = ((RecentFileInfo) RecentFileListBox.SelectedItem).File;
             _recentFileList.RemoveFile(file);
         }
